+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * (generic) `DOS _MKDIR`、`DOS _RMDIR`、`DOS _CHDIR`、`DOS _FILEDATE`を実装。
+* (win32) `DOS _FILEDATE`で予約デバイス名は取得・設定せず常に0を返すようにした。
 
 
 ## 2.2.0 (2025-10-30)